pyproject.toml

[project]
name = "v26meme"
version = "0.1.0"
description = "Autonomous, self-improving crypto trading intelligence (paper/live-optional)"
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.26",
    "pandas>=2.2",
    "pydantic>=2.7",
    "scipy>=1.11",
    "ccxt>=4.3.69",
    "plotly>=5.20",
    "streamlit>=1.35",
    "pyyaml>=6.0.1",
    "click>=8.1.7"
]

[build-system]
requires = ["setuptools>=67", "wheel"]
build-backend = "setuptools.build_meta"
